Extract server API key route helpers
